mito-ai-cli: strip mito citationa and cell references
